% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.r
\name{compile.package}
\alias{compile.package}
\title{compile.package}
\usage{
\code{compile.package(path=STRING, as.test=BOOL)}
}
\arguments{
\item{path}{path name of package.}

\item{as.test}{default to \code{FALSE}. If set to \code{TRUE}, will install package locally without compilation. If set to \code{FALSE}/unset will compile via roxygen.}
}
\description{
Compiles R package using roxygen2.
}
\examples{
\dontrun{
compile.package(path='~/Documents/R/packages/mytestpackage');
compile.package(path='~/Documents/R/packages/mytestpackage', as.test=TRUE);
};

}
\keyword{compile}
\keyword{package}
\keyword{syntax}
