% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.r
\name{console.clear}
\alias{console.clear}
\title{console.clear}
\usage{
\code{console.clear()}
}
\description{
Allow for simpler syntax in R. Clear console.
}
\examples{
\dontrun{
console.clear();
}

}
\keyword{console.clear}
\keyword{syntax}
