% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.from.url.r
\name{install.from.url}
\alias{install.from.url}
\title{rbettersyntax | install.from.url}
\usage{
install.from.url(url = NULL, unzip = FALSE, install = TRUE)
}
\arguments{
\item{url}{a character string. A url.}

\item{unzip}{boolean. Default \code{FALSE}. If set to \code{TRUE}, then will the downloaded object will be unpacked.}

\item{install}{boolean. Default \code{TRUE}. If set to \code{TRUE}, then the package in the downloaded object will be installed, otherwise the path of the downloaded package will be returned.}
}
\description{
Allow for simpler syntax in R. Loads packages from URL, potentially unzipping if necessary. By contrast to \code{installr}, the URLs with parameters do not lead to problems.
}
\details{
\code{install.from.url(url='...', unzip=TRUE/FALSE, install=TRUE/FALSE)}
}
\examples{
rbettersyntax::install.from.url(url='http://domain.de/mypackage.zip?parameters=areok', unzip=TRUE);
rbettersyntax::install.from.url(url='http://domain.de/mypackage.gz?parameters=areok', unzip=TRUE);
rbettersyntax::install.from.url(url='http://domain.de/mypackage?parameters=areok');
path <- rbettersyntax::install.from.url(url='http://domain.de/mypackage?parameters=areok', install=FALSE);
install.packages(pkgs=path, repos=NULL, type='source');

}
\keyword{URL}
\keyword{install}
\keyword{load}
\keyword{packages}
\keyword{syntax}
