% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.r
\name{load.packages}
\alias{load.packages}
\title{load.packages}
\usage{
\code{load.packages(mode=STRING, package=STRING/s, url=STRING/s, mirror=INT, force=BOOL, dependencies=BOOL/STRINGS, quietly=BOOL, stoponerror=BOOL, lib=STRING, file.type=STRING, version=STRING)}
}
\arguments{
\item{mode}{one of \code{'cran'} (Default), \code{'url'}, \code{'github'} or \code{'biocmanager'}. Determines how the package will be loaded. If \code{mode='url'} is set, then there setting the \code{package} vector is optional.}

\item{package}{a vector of strings.}

\item{url}{a vector of strings. Either \code{package} or \code{url} should be non-empty.}

\item{mirror}{an integer between 0 and (currently) 94. If not set, will ask for user input, upon the first time that CRAN-sources are loaded.}

\item{file.type}{an optional string, e.g. \code{'zip'} or \code{'gz'}, to be use under \code{mode='url'}, if the downloaded file is to be first unzipped.}

\item{lib}{an optional string, the Git-repository name.}

\item{dependencies}{either \code{TRUE}, \code{FALSE} or a string-vector. A string-vector is only useful in the case of installing under \code{mode='url'}, and in this one may use either boolean values or a subset of \code{c('Depends', 'Imports', 'LinkingTo', 'Suggests', 'Enhances')}.}

\item{force}{if set \code{TRUE}, will force an installation. If \code{FALSE} (Default), will first attempt to load the package and then install.}

\item{quietly}{boolean value to control the verbosity of installers.}

\item{stoponerror}{a boolean value (default \code{FALSE}). If set as \code{TRUE}, will stop, if loading any package (after attempting to install it) fails.}

\item{version}{a string for package versioning. Useful only under \code{mode='biocmanager'}.}
}
\description{
Loads packages. If not available, then an install attempt will be carried out.
}
\examples{
\dontrun{
load.packages(mirror=53, package='tidyverse', force=TRUE);
load.packages(mirror=53, package=c('tidyverse', 'Matrix'), stoponerror=TRUE, force=TRUE);
load.packages(mirror=53, package=c('tidyverse', 'Motrix', 'cowplot'), stoponerror=TRUE, force=TRUE); # this will stop!
load.packages(mode='github', package=c('RLogik/clusterby', 'RLogik/rbettersyntax'), dependencies=TRUE);
load.packages(mode='github', url='https://github.com/RLogik/clusterby/archive/master.zip', file.type='zip', force=TRUE, dependencies=FALSE);

## require('clusterby') will be tried first, then potentially an install, then a require.
load.packages(mode='github', package='clusterby', url='https://github.com/RLogik/clusterby/archive/master.zip', file.type='zip', dependencies=TRUE);
};

}
\keyword{install}
\keyword{load}
\keyword{packages}
\keyword{syntax}
