% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ok.comma.r
\name{ok.comma}
\alias{ok.comma}
\title{rbettersyntax | ok.comma}
\usage{
\code{f <- ok.comma(f)}
}
\arguments{
\item{f}{Function to be modified to allow trailing commas}
}
\description{
Allow for simpler syntax in R. This method enables trailing commata. Modification of flodel's method.
}
\examples{
\dontrun{
c <- rbettersyntax::ok.comma(base::c);
list <- rbettersyntax::ok.comma(base::list);
myfun <- rbettersyntax::ok.comma(myfun);
}

}
\keyword{comma}
\keyword{syntax}
\keyword{trailing}
