% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.r
\name{read.args}
\alias{read.args}
\title{read.args}
\usage{
\code{val <- read.args(args, key, type, defaultval)}
}
\arguments{
\item{key}{character string. This is the name of the parameter you wish to extract.}

\item{type}{function or string. If a boolean-valued function like \code{is.character}, \code{is.logical}, etc., this will be used to check the type of the argument. If a string, then a standard type-checking function will be called.}

\item{args}{Liste of arguments extracted from ... . Usage: inside a function first set \code{args <- list(...);}.}

\item{defaultval}{Default value. Will be set, if argument is missing or is not of the correct type.}
}
\description{
Read input arguments from ... with expected type and default values.
}
\examples{
\dontrun{
f <- function(...) {
	args <- list(...);
	val <- read.args(args, key='name', type=is.character, default='N/A');
	## Rest of the function definition
};
 }

}
\keyword{arguments}
\keyword{default}
\keyword{parameters}
\keyword{read}
\keyword{syntax}
\keyword{value}
