% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.multiple.r
\name{return.multiple}
\alias{return.multiple}
\title{rbettersyntax | return.multiple}
\usage{
\code{return.multiple(map, f)(<INPUTARGS>)}
}
\arguments{
\item{map}{A list mapping external names to local-names of return-entities in output of f.}

\item{f}{Function which returns a list-objet as output.}

\item{<INPUTARGS>}{Input arguments for f.}
}
\description{
Allow for simpler syntax in R. Handles functions with multiple outputs better without messy intermediate list-objects.
}
\examples{
\dontrun{
infos <- function(firstname='',surname='',age=0) {
	underage <- (age < 18);
	fullname <- gsub('^\\\\s+|\\\\s+$', '', print0(firstname,' ',surname));

	return(list(
		voter = !underage,
		fullname = fullname,
		validname = !(fullname == '')
	));
};
return.multiple(list(valid='validname', isvoter='voter'), f)(firstname='Richard', surname='Feynman', age=100);

# equivalent to:
obj <- f(firstname='Richard', surname='Feynman', age=100);
valid <- obj$validname;
isvoter <- obj$voter;
 }
\dontrun{
f <- function(<INPUTARGS>) {
	·
	·
	·
	return(list(
		localoutput1 = <RETURN-OUTPUT1>,
		localoutput2 = <RETURN-OUTPUT2>,
		·
		·
	));
};
 ·
 ·
 ·
return.multiple(list(var1=localoutput1, var2=localoutput2, ...), f)(<INPUTARGS>);

# äquivalent zu:
obj <- f(<INPUTARGS>);
var1 <- obj$localoutput1;
var2 <- obj$localoutput1;
 ·
 ·
 ·
 }

}
\keyword{arguments}
\keyword{multiple}
\keyword{return}
\keyword{syntax}
