% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.setup.r
\name{standard.setup}
\alias{standard.setup}
\title{rbettersyntax | standard.setup}
\usage{
standard.setup()
}
\description{
Provides standard ok.comma'd functions and enables via option a disabling of menu and sys.pause.
}
\details{
\code{ok.comma.standard.setup()}
}
\examples{
options('rbettersyntax::silent'=TRUE);
options('rbettersyntax::silent'=FALSE);
rbettersyntax::standard.setup();

}
\keyword{rbettersyntax}
\keyword{setup}
\keyword{standard}
\keyword{syntax}
