% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.pause.r
\name{sys.pause}
\alias{sys.pause}
\title{rbettersyntax | sys.pause}
\usage{
sys.pause(t = NULL)
}
\arguments{
\item{t}{non-negative double. Argument is optional. If \code{t} is set, then the script will be paused for \code{t} seconds. If it is not set, then a user key-response will be awaited.}
}
\description{
Allow for simpler syntax in R. Method to pause execution. Use \code{options('rbettersyntax::silent'=TRUE)} to disable all sys.pause commands in e. g. markdown mode.

\code{sys.pause(* t)}
}
\examples{
\dontrun{
sys.pause(); # pause until user presses any key.
sys.pause(0.5); # pause for 0.5 seconds
sys.pause(10); # pause for 10 seconds
}

}
\keyword{pause}
\keyword{syntax}
\keyword{system}
