% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{this.dir}
\alias{this.dir}
\title{this.dir}
\usage{
\code{this.dir(chdir=<BOOL>);}
}
\arguments{
\item{chdir}{boolean. Default \code{chdir=TRUE}. Whether or not to change directory to location of current script, or simply return the path.}
}
\description{
Change the current directory (for RStudio).
}
\examples{
\dontrun{
# Prefix script with:
bool <- this.dir();
# or
bool <- this.dir(chdir=TRUE);
# to change to current directory.
# Return value: bool = TRUE <==> succesfully found currently and changed to this path.

path <- this.dir(chdir=TRUE);
# does not change directory, but attempts to get
# the folder location of current script.
# If unsuccessful, returns path = NULL.
}

}
\keyword{RStudio}
\keyword{change}
\keyword{direcctory}
\keyword{syntax}
