% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.r
\docType{class}
\name{timer-class}
\alias{timer-class}
\alias{timer}
\title{timer}
\usage{
\code{clock <- timer()}
}
\description{
Method to pause execution. Use \code{options('rbettersyntax::silent'=TRUE)} to disable all sys.pause commands in e. g. markdown mode.
}

\examples{
\dontrun{
clock <- timer();

clock$time.absolute(); # return absolute time in seconds.
clock$time(); # return relative time since last reset.
clock$duration(); # return relative time since last stop.
clock$reset(); # resets start time.
clock$stop(); # preserves start time, but starts a new lap.

clock$pause(); # pause until user presses any key.
clock$pause(0.5); # pause for 0.5 seconds
clock$pause(10); # pause for 10 seconds
}

}
\keyword{pause}
\keyword{syntax}
\keyword{time}
