% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.r
\name{trailing.comma}
\alias{trailing.comma}
\title{trailing.comma}
\usage{
\code{f <- trailing.comma(f)}
}
\arguments{
\item{f}{Function to be modified to allow trailing commas}
}
\description{
This method enables trailing commata. Modification of flodel's method.
}
\examples{
\dontrun{
c <- trailing.comma(base::c);
list <- trailing.comma(base::list);
myfun <- trailing.comma(myfun);
}

}
\keyword{comma}
\keyword{syntax}
\keyword{trailing}
