% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.clear.r
\name{console.clear}
\alias{console.clear}
\title{rbettersyntax | console.clear}
\usage{
console.clear()
}
\description{
Allow for simpler syntax in R. Clear console.

\code{console.clear()}
}
\examples{
\dontrun{
console.clear();
}

}
\keyword{console.clear}
\keyword{syntax}
