% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.log.r
\name{console.log}
\alias{console.log}
\title{rbettersyntax | console.log}
\usage{
console.log(tabs = 0, ...)
}
\arguments{
\item{tabs}{a non-negative integer. Number of tab sets to be produced on each new line.}

\item{m1}{character string / vector of character strings. The strings will be concatenated with no spaces, prepended by tab stops and followed by a \code{'\n'} character.}

\item{m2}{similar. User can set 0 or more 'messages' to be printed in the console.}
}
\description{
Allow for simpler syntax in R. Output messages to console
}
\details{
\code{console.log(tabs, m1, m2, ..)}
}
\examples{
console.log(0, 'Test');
console.log(1, 'Starting Method', c('Scanning data set ',k,' for solutions:'));

}
\keyword{console}
\keyword{console.log}
\keyword{log}
\keyword{syntax}
