% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.r
\name{console.log}
\alias{console.log}
\title{console.log}
\usage{
\code{console.log(tabs=tabs, tab.char=tab.char, silent.off=silent.off, m1, m2, ..)}
}
\arguments{
\item{tabs}{a non-negative integer. Default \code{0}. Number of tab sets to be produced on each new line.}

\item{tab.char}{character. Default \code{'  '}. This will be used as the indentation string.}

\item{silent.off}{boolean. Default \code{FALSE}. If true, then no output will be shown under the option \code{'utilsrl::silent'=TRUE}.}

\item{m1}{character string / vector of character strings. The strings will be concatenated with no spaces, prepended by tab stops and followed by a \code{'\n'} character.}

\item{m2}{similar. User can set 0 or more 'messages' to be printed in the console.}
}
\description{
Allow for simpler syntax in R. Output messages to console. Use \code{options('utilsrl::silent'=TRUE)} to disable console output.
}
\examples{
\dontrun{
console.log(tabs=0, 'Test');
console.log(tabs=1, silent.off=TRUE, 'Starting Method', c('Scanning data set ',k,' for solutions:'));
}

}
\keyword{console}
\keyword{console.log}
\keyword{log}
\keyword{syntax}
