% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.r
\name{install.from.url}
\alias{install.from.url}
\title{install.from.url}
\usage{
\code{install.from.url(pkg.name='...', url='...', file.type='...', install=TRUE/FALSE, require.pkg=TRUE/FALSE, force=TRUE/FALSE)}
}
\arguments{
\item{pkg.name}{a character string. Optional. Name of Package, if known. Otherwise, the proceedure extracts this from the downloaded file.}

\item{url}{a character string. A url.}

\item{file.type}{character. Default \code{NULL}. If not set, then system will attempt to determin whether the file is to be extracted and by what method. If set to \code{'zip'}, \code{'gz'}, \code{'7z'}, etc., then an appropriate method will be applied to unpack the downloaded object.}

\item{install}{boolean. Default \code{TRUE}. If set to \code{TRUE}, then the package in the downloaded object will be installed, otherwise the path of the downloaded package will be returned.}

\item{require.pkg}{boolean. Default \code{FALSE}. If \code{require.pkg=TRUE} and \code{force=FALSE} and \code{pkg.name} is defined, \code{require(pkg.name)} will be called at the start. If it succeeds, no installation will take place. If \code{require.pkg=TRUE}, then \code{require(···)} will be called at the end.}

\item{force}{boolean. Defaut \code{FALSE}. If set to \code{FALSE}, determines, whether \code{library(···)} / \code{require(···)} will be attempted first before installation.}
}
\description{
Allow for simpler syntax in R. Loads packages from URL, potentially unzipping if necessary. By contrast to \code{installr}, the URLs with parameters do not lead to problems.
}
\examples{
\dontrun{
install.from.url(url='http://domain.de/mypackage.zip?parameters=areok', file.type='zip');
install.from.url(url='http://domain.de/mypackage.gz?parameters=areok', file.type='gz');
install.from.url(url='http://domain.de/mypackage.gz?parameters=areok'); ## file type will be automatically detected
install.from.url(url='http://domain.de/mypackage?parameters=areok');
path <- install.from.url(url='http://domain.de/mypackage?parameters=areok', install=FALSE);
install.packages(pkgs=path, repos=NULL, type='source');
path <- install.from.url(url='http://domain.de/mypackage?parameters=areok', install=FALSE, pkg.name='mypackage', require.pkg=TRUE); ## tries require('mypackage') first, otherwise installs package afresh.
path <- install.from.url(url='http://domain.de/mypackage?parameters=areok', install=FALSE, require.pkg=TRUE, force=TRUE); ## forces a fresh installation of the package followed by a require(···) command (package name will be automatically detected).
}

}
\keyword{URL}
\keyword{install}
\keyword{load}
\keyword{packages}
\keyword{syntax}
