% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.args.r
\name{read.args}
\alias{read.args}
\title{rbettersyntax | read.args}
\usage{
read.args(vars, ...)
}
\arguments{
\item{args}{Liste of arguments extracted from ... . Usage: inside a function first set \code{args <- list(...);}.}

\item{key}{character string. This is the name of the parameter you wish to extract.}

\item{type}{function or string. If a boolean-valued function like \code{is.character}, \code{is.logical}, etc., this will be used to check the type of the argument. If a string, then a standard type-checking function will be called.}

\item{defaultval}{Default value. Will be set, if argument is missing or is not of the correct type.}
}
\description{
Allow for simpler syntax in R. Read input arguments from ... with expected type and default values.
}
\details{
\code{val <- read.args(args, key, type, defaultval)}
}
\keyword{arguments}
\keyword{default}
\keyword{parameters}
\keyword{read}
\keyword{syntax}
\keyword{value}
