% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{standard.setup}
\alias{standard.setup}
\title{standard.setup}
\usage{
\code{standard.setup();}
}
\arguments{
\item{options}{character string vector. As per the definition of base::menu}
}
\description{
Provides standard trailing.comma'd functions and enables via option a disabling of menu and sys.pause.
}
\examples{
\dontrun{
options('utilsrl::silent'=TRUE);
standard.setup();
# will skip user interaction and return \\code{0}:
menu(c('Yes','No'), title='Proceed with script?');
sys.pause(); # will skip user interaction proceed with code.
sys.pause(1); # will still pause for 1s.
console.log(tabs=0, 'My message.'); # will not display.
console.log(tabs=0, silent.off=TRUE, 'My message.'); # will not display.
console.log(tabs=0, silent.off=FALSE, 'My message.'); # will show.
# reactivates menu, sys.pause, console.log, etc.
options('utilsrl::silent'=FALSE);
}

}
\keyword{rbettersyntax}
\keyword{setup}
\keyword{standard}
\keyword{syntax}
